test(build-lease): split cap-arming coverage into its own module
`Server Guards` rejected `build_lease_integration_tests.rs` at 52985 bytes
against the 51200-byte ceiling (the line count, 1464, was still under 1500).
The new cap-arming tests are a self-contained concern with their own setup,
so they move to `build_lease_cap_arming_tests.rs` rather than earning the
file a `djinn:allow-oversize` marker — the guard is asking for a split, and a
marker here would only defer it.

Pure code motion plus the module registration. Both files now measure under
the line and byte ceilings.
